TES-7: Integrate OpenAI API to Generate Ad Headlines
