+++ app/main.py
@@ -7,6 +7,7 @@ import requests
 from bs4 import BeautifulSoup
 import os
 from app.scraping import extract_images
+from app.openai_utils import generate_headline  # Import the function here
 from dotenv import load_dotenv
 import openai
 
@@ -27,20 +28,6 @@ templates = Jinja2Templates(directory="templates")
 @app.get("/", response_class=HTMLResponse)
 def read_root(request: Request):
     return templates.TemplateResponse("index.html", {"request": request})
-
-
-def generate_headline(text: str, image_url: str) -> str:
-    prompt = f"Generate an ad headline for the following image URL based on the website text:\n\nWebsite Text:\n{text}\n\nImage URL: {image_url}\n\nAd Headline:"
-    response = openai.ChatCompletion.create(
-        model="gpt-3.5-turbo",
-        messages=[
-            {"role": "system", "content": "You are a helpful assistant."},
-            {"role": "user", "content": prompt}
-        ],
-        max_tokens=50
-    )
-    return response.choices[0].message['content'].strip()
-
 
 @app.get("/extract-text")
 def extract_text(url: HttpUrl = Query(..., description="The URL to extract text and images from")):
